Update OS X build options for Mathematica 10.4.1
It appears that newer versions of Mathematica have two versions of the
MathLink library available, one which uses libc++ and another which uses
the older libstdc++. Only the latter supports older versions of OS X as
far back as OS X 10.7 (Lion), so we need to explicitly used that one.

This change also means that h5mma now requires OS X 10.7 as a minimum,
rather than the previous 10.6 requirement.